Reorganize Media related files into consolidated Media module
