+++ include/BLIB/Media.hpp
@@ -3,11 +3,29 @@
 
 /**
  * @defgroup Media
- * @brief Collection of utility classes extending SFML's base rendering capabilities
- * 
+ * @brief Provides several groups of classes for more advanced media functionality
+ *
  */
 
-#include <BLIB/Media/AnimationData.hpp>
-#include <BLIB/Media/Animation.hpp>
+/**
+ * @addtogroup Graphics
+ * @brief Provides several different graphics features such as animations
+ * @ingroup Media
+ *
+ */
+
+/**
+ * @addtogroup Shapes
+ * @brief Collection of shapes to add to the functionality provided by SFML
+ * @ingroup Media
+ *
+ */
+
+#include <BLIB/Media/Graphics/Animation.hpp>
+#include <BLIB/Media/Graphics/AnimationData.hpp>
+
+#include <BLIB/Media/Shapes/Ellipse.hpp>
+#include <BLIB/Media/Shapes/GradientCircle.hpp>
+#include <BLIB/Media/Shapes/Triangle.hpp>
 
 #endif
