+++ include/BLIB/Media/Graphics/AnimationData.hpp
