+++ include/BLIB/Media/Graphics/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(BLIB PUBLIC
+    AnimationData.hpp
+    Animation.hpp
+)
