+++ include/BLIB/Media/Shapes/GradientCircle.hpp
