+++ include/BLIB/Media/Shapes/Triangle.hpp
