--- include/BLIB/Shapes.hpp
@@ -1,14 +1,0 @@
-#ifndef BLIB_SHAPES_HPP
-#define BLIB_SHAPES_HPP
-
-/**
- * @defgroup Shapes
- * @brief Collection of shapes to add to the functionality provided by SFML
- *
- */
-
-#include <BLIB/Shapes/Ellipse.hpp>
-#include <BLIB/Shapes/GradientCircle.hpp>
-#include <BLIB/Shapes/Triangle.hpp>
-
-#endif
