+++ src/CMakeLists.txt
@@ -9,5 +9,4 @@ add_subdirectory(Media)
 add_subdirectory(Menu)
 add_subdirectory(Parser)
 add_subdirectory(Scripts)
-add_subdirectory(Shapes)
 add_subdirectory(Util)
