+++ src/Media/Graphics/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(BLIB PRIVATE
+    AnimationData.cpp
+    Animation.cpp
+)
